+++ src/internal/data.ts
@@ -1,4 +1,4 @@
-import type { Model, Schema, SchemaDefinition } from 'mongoose';
+import type { Model, SchemaDefinition } from 'mongoose';
 
 import type { IGlobalOptions, VirtualOptions } from '../types';
 import { Severity } from './constants';
@@ -17,10 +17,6 @@ export interface IHooks {
   pre: IPreHook[];
   post: IPostHook[];
 }
-export interface IPluginMap {
-  mongoosePlugin(schema: Schema<any>, options: object): void;
-  options: any;
-}
 
 /** Schema Map */
 export const schemas: Map<string, SchemaDefinition> = new Map();
@@ -30,8 +26,6 @@ export const models: Map<string, Model<any>> = new Map();
 export const virtuals: Map<string, Map<string, VirtualOptions>> = new Map();
 /** Hooks Map */
 export const hooks: Map<string, IHooks> = new Map();
-/** Plugins Map */
-export const plugins: Map<string, IPluginMap[]> = new Map();
 /** Constructors Map */
 export const constructors: Map<string, NewableFunction> = new Map();
 /** Global Options */